decimal changes

+++ PrototypeWithAuth/Models/ExchangeRate.cs
@@ -8,7 +8,7 @@ namespace PrototypeWithAuth.Models
     public class ExchangeRate
     {
         public int ExchangeRateID { get; set; }
-        public Double LatestExchangeRate { get; set; }
+        public decimal LatestExchangeRate { get; set; }
         public DateTime LastUpdated { get; set; }
     }
 }
